Remove homedir fetching
...due to it being unsupported in the scratch image without copying in
/etc/passwd, which I'd really prefer not to do.

The reason this isn't supported is because the scratch image has no
/etc/passwd, we've compiled without cgo (allows using standard
libraries), and we're essentially running in a user-less environment.

I the future, I'll try to find a way to detect this state, but for now,
it's not really necessary. If someone wants to run with a config file in
their homedir, they can use the -config option.